Amended default display of SVG element
SVG defaults to display: inline which I dislike. Previously had this set to display: inline-flex but inline-block or inline-flex will render white space on SVG elements in HTML (when in body element where you would have defs and symbols) if the container isn't a flex box or the font-size set to 0 to crush the whitespace.